Moving events to constants subdir
